Add files via upload
model

The model used is a Fuzzy Logic system that receives data in the form of crisp values. These values are processed in a Membership function, which has three levels (Low, Medium, High). Afterward, rules are applied to determine the risk level."
Dataset

The dataset type is simulation data, representing virtual readings from sensors. Its size is not fixed as it depends on the number of times the program is run.