Add files via upload
Model

The model used is a (Random Forest Classifier)implemented within a pipeline. This pipeline includes steps for data standardization using (StandardScaler) and class balancing with (SMOTE) to address class imbalance. The (Random Forest) algorithm constructs multiple decision trees and aggregates their results for accurate predictions. (GridSearchCV) is employed to fine-tune hyperparameters like the number of trees and maximum tree depth.
Dataset

The dataset is simulated data with 1000 samples and 7 features, representing sensor readings. The target variable, alert, is binary and imbalanced with 90% of samples in one class. The data is split into 70% for training and 30% for testing.